@@ -1,6 +1,6 @@
 package org.sanjay.javabrains.messanger.model;
 
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class ErrorMessage {
