@@ -2,7 +2,7 @@ package org.sanjay.javabrains.messanger.model;
 
 import java.util.Date;
 
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class Profile {
